refactor : Use `strongunits.MiB` / `strongunits.GiB` to store memory size (#1642)
Currently we're using memory values in `uint64` variable. This works but sometimes it can be confusing to figure out what unit we're using if variable is not named correctly.

Instead, we can rely on strongunits types for byte,MegaByte and GigaByte for these values.

+++ pkg/crc/config/settings.go
@@ -156,7 +156,7 @@ func defaultCPUs(cfg Storage) uint {
 }
 
 func defaultMemory(cfg Storage) uint {
-	return constants.GetDefaultMemory(GetPreset(cfg))
+	return uint(constants.GetDefaultMemory(GetPreset(cfg)))
 }
 
 func defaultBundlePath(cfg Storage) string {
